refactor(llm): isolate NIXL media transport
