+++ lib/llm/src/kv_router/indexer/embedding_cache.rs
@@ -40,6 +40,7 @@ pub fn preprocessed_multimodal_cache_keys(request: &PreprocessedRequest) -> Vec<
         match item {
             MultimodalData::Url(url) => keys.push(multimodal_cache_key_from_url(url.as_str())),
             MultimodalData::RawUrl(url) => keys.push(multimodal_cache_key_from_url(url)),
+            #[cfg(feature = "media-nixl")]
             MultimodalData::Decoded(descriptor) => {
                 if let Some(key) = descriptor.content_hash_key() {
                     keys.push(key.to_string());
