+++ lib/llm/src/preprocessor/media.rs
@@ -2,38 +2,25 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod common;
+mod decoded;
 mod decoders;
 mod jpeg_turbo;
 mod loader;
-mod rdma;
-
 #[cfg(feature = "media-nixl")]
-use anyhow::{Context, Result};
-#[cfg(feature = "media-nixl")]
-use dynamo_protocols::types::ChatCompletionRequestMessageContentPartImage;
+mod nixl;
 
 pub use common::EncodedMediaData;
+pub use decoded::DecodedMediaData;
 pub use decoders::{Decoder, ImageDecoder, MediaDecoder};
-pub use loader::{MediaFetcher, MediaLoader};
-
-pub use rdma::{DecodedMediaData, RdmaMediaDataDescriptor};
+pub use loader::MediaFetcher;
 #[cfg(feature = "media-nixl")]
-pub use rdma::{get_nixl_agent, get_nixl_metadata};
+pub use nixl::{MediaLoader, RdmaMediaDataDescriptor, get_nixl_agent, get_nixl_metadata};
+
+/// Marker used internally when frontend media decoding is not compiled in.
+#[cfg(not(feature = "media-nixl"))]
+pub struct MediaLoader;
 
 #[doc(hidden)]
 pub fn libjpeg_turbo_available() -> bool {
     jpeg_turbo::available()
-}
-
-#[cfg(feature = "media-nixl")]
-pub(super) fn require_image_url(
-    part: &ChatCompletionRequestMessageContentPartImage,
-) -> Result<&url::Url> {
-    Ok(&part
-        .image_url
-        .as_ref()
-        .context(
-            "Cannot decode an image content part without a URL; UUID-only parts must be resolved by the backend cache",
-        )?
-        .url)
 }
